==> notebooks/discovery/1.2.discovery_data_type.ipynb <==
import pandas as pd

# Chargement du fichier avec le bon séparateur
df = pd.read_csv("../../data/raw/Classeur1.csv", sep=";")

# Fonction pour compter les types de valeur
def count_value_types(series):
    return series.dropna().apply(lambda x: type(x).__name__).value_counts()

# Appliquer à toutes les colonnes et uniformiser les clés
type_counts = {}
for col in df.columns:
    counts = count_value_types(df[col])
    counts.index = [str(t) for t in counts.index]  # Convertir les types en chaînes
    type_counts[col] = counts

# Créer un DataFrame à partir du dict, remplacement NaN par 0
type_df = pd.DataFrame.from_dict(type_counts, orient='index').fillna(0).astype(int)

# Afficher le résultat
display(type_df)

# Chargement du fichier avec le bon séparateur
df = pd.read_csv("../../data/cleaning/step1_typed.csv")

# Fonction pour compter les types de valeur
def count_value_types(series):
    return series.dropna().apply(lambda x: type(x).__name__).value_counts()

# Appliquer à toutes les colonnes et uniformiser les clés
type_counts = {}
for col in df.columns:
    counts = count_value_types(df[col])
    counts.index = [str(t) for t in counts.index]  # Convertir les types en chaînes
    type_counts[col] = counts

# Créer un DataFrame à partir du dict, remplacement NaN par 0
type_df = pd.DataFrame.from_dict(type_counts, orient='index').fillna(0).astype(int)

# Afficher le résultat
display(type_df)
